pyproject.toml
[tool.poetry]
name = "book-gen-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^3.0.2"
openpyxl = "^3.1.5"
supabase = "^2.28.3"
python-dotenv = "^1.2.2"
openai = "^2.30.0"
python-docx = "^1.2.0"
reportlab = "^4.4.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
